@@ -1,75 +1,86 @@
 {
   "private": true,
   "engines": {
-    "node": ">= 0.10",
-    "npm": ">= 2.1"
+    "node": ">=5.0 <6",
+    "npm": ">=3.3 <4"
   },
   "dependencies": {
-    "babel": "5.8.21",
-    "classnames": "2.1.3",
+    "babel-core": "5.8.34",
+    "bluebird": "3.1.1",
+    "classnames": "2.2.1",
     "eventemitter3": "1.1.1",
     "express": "4.13.3",
     "fastclick": "1.0.6",
-    "fbjs": "0.1.0-alpha.7",
-    "flux": "2.0.3",
-    "front-matter": "1.0.0",
+    "fbjs": "0.5.1",
+    "front-matter": "2.0.1",
+    "history": "1.16.0",
     "jade": "1.11.0",
-    "lodash": "3.10.1",
+    "node-fetch": "1.3.3",
     "normalize.css": "3.0.3",
-    "react": "0.14.0-beta3",
-    "react-dom": "0.14.0-beta3",
-    "react-routing": "0.0.4",
-    "source-map-support": "0.3.2",
-    "superagent": "1.3.0"
+    "react": "0.14.3",
+    "react-dom": "0.14.3",
+    "react-routing": "0.0.6",
+    "source-map-support": "0.4.0",
+    "whatwg-fetch": "0.10.1"
   },
   "devDependencies": {
-    "autoprefixer-core": "^5.2.1",
-    "babel-eslint": "^4.0.5",
-    "babel-loader": "^5.3.2",
-    "browser-sync": "^2.8.2",
-    "css-loader": "^0.15.6",
+    "assets-webpack-plugin": "^3.2.0",
+    "autoprefixer": "^6.1.2",
+    "babel": "^5.8.34",
+    "babel-eslint": "^4.1.6",
+    "babel-loader": "^5.4.0",
+    "babel-plugin-react-transform": "^1.1.1",
+    "browser-sync": "^2.10.1",
+    "css-loader": "^0.23.0",
     "csscomb": "^3.1.8",
-    "cssnext": "^1.8.3",
-    "del": "^1.2.0",
-    "eslint": "^1.1.0",
-    "eslint-loader": "^1.0.0",
-    "eslint-plugin-react": "^3.2.2",
-    "gaze": "^0.5.1",
-    "git-push": "^0.1.1",
-    "glob": "^5.0.14",
-    "jest-cli": "^0.4.18",
+    "del": "^2.2.0",
+    "eslint": "^1.10.3",
+    "eslint-config-airbnb": "^2.1.1",
+    "eslint-loader": "^1.1.1",
+    "eslint-plugin-react": "^3.11.3",
+    "file-loader": "^0.8.5",
+    "gaze": "^0.5.2",
+    "git-repository": "^0.1.1",
+    "glob": "^6.0.1",
+    "isomorphic-style-loader": "0.0.5",
+    "jest-cli": "^0.8.2",
+    "jscs": "^2.7.0",
+    "json-loader": "^0.5.4",
+    "lodash.merge": "^3.3.2",
     "mkdirp": "^0.5.1",
     "ncp": "^2.0.0",
-    "postcss": "^4.1.16",
-    "postcss-loader": "^0.5.1",
-    "postcss-nested": "^0.3.2",
-    "psi": "^1.0.6",
-    "react-hot-loader": "^1.2.8",
+    "postcss": "^5.0.13",
+    "postcss-import": "^7.1.3",
+    "postcss-loader": "^0.8.0",
+    "precss": "^1.3.0",
+    "raw-loader": "^0.5.1",
+    "react-transform-catch-errors": "^1.0.0",
+    "react-transform-hmr": "^1.0.1",
+    "redbox-react": "^1.2.0",
     "replace": "^0.3.0",
-    "run-sequence": "^1.1.2",
-    "style-loader": "^0.12.3",
-    "url-loader": "^0.5.6",
-    "webpack": "^1.11.0",
-    "webpack-dev-middleware": "^1.2.0",
-    "webpack-hot-middleware": "^1.2.0"
+    "url-loader": "^0.5.7",
+    "webpack": "^1.12.9",
+    "webpack-hot-middleware": "^2.6.0",
+    "webpack-middleware": "^1.4.0"
   },
   "jest": {
     "rootDir": "./src",
     "scriptPreprocessor": "../preprocessor.js",
     "unmockedModulePathPatterns": [
+      "fbjs",
       "react"
     ]
   },
   "scripts": {
-    "lint": "eslint src tools",
+    "lint": "eslint src tools && jscs src tools",
     "csslint": "csscomb src/components --lint --verbose",
     "csscomb": "csscomb src/components --verbose",
     "test": "eslint src && jest",
-    "clean": "babel-node --eval \"require('./tools/clean')().catch(err => console.log(err.stack))\"",
-    "copy": "babel-node --eval \"require('./tools/copy')().catch(err => console.log(err.stack))\"",
-    "bundle": "babel-node --eval \"require('./tools/bundle')().catch(err => console.log(err.stack))\"",
-    "build": "babel-node --eval \"require('./tools/build')().catch(err => console.log(err.stack))\"",
-    "serve": "babel-node --eval \"require('./tools/serve')().catch(err => console.log(err.stack))\"",
-    "start": "babel-node --eval \"require('./tools/start')().catch(err => console.log(err.stack))\""
+    "clean": "babel-node tools/run clean",
+    "copy": "babel-node tools/run copy",
+    "bundle": "babel-node tools/run bundle",
+    "build": "babel-node tools/run build",
+    "deploy": "babel-node tools/run deploy",
+    "start": "babel-node tools/run start"
   }
 }
