@@ -1,42 +1,54 @@
-/*! React Starter Kit | MIT License | http://www.reactstarterkit.com/ */
+/**
+ * React Starter Kit (https://www.reactstarterkit.com/)
+ *
+ * Copyright © 2014-2016 Kriasoft, LLC. All rights reserved.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE.txt file in the root directory of this source tree.
+ */
 
+import fs from 'fs';
 import { join } from 'path';
 import { Router } from 'express';
+import Promise from 'bluebird';
 import jade from 'jade';
 import fm from 'front-matter';
-import fs from '../utils/fs';
 
 // A folder with Jade/Markdown/HTML content pages
 const CONTENT_DIR = join(__dirname, './content');
 
 // Extract 'front matter' metadata and generate HTML
 const parseJade = (path, jadeContent) => {
-  const content = fm(jadeContent);
-  const html = jade.render(content.body, null, '  ');
-  const page = Object.assign({ path, content: html }, content.attributes);
-  return page;
+  const fmContent = fm(jadeContent);
+  const htmlContent = jade.render(fmContent.body);
+  return Object.assign({ path, content: htmlContent }, fmContent.attributes);
 };
+
+const readFile = Promise.promisify(fs.readFile);
+const fileExists = filename => new Promise(resolve => {
+  fs.exists(filename, resolve);
+});
 
 const router = new Router();
 
 router.get('/', async (req, res, next) => {
   try {
-    let path = req.query.path;
+    const path = req.query.path;
 
     if (!path || path === 'undefined') {
-      res.status(400).send({error: `The 'path' query parameter cannot be empty.`});
+      res.status(400).send({ error: `The 'path' query parameter cannot be empty.` });
       return;
     }
 
     let fileName = join(CONTENT_DIR, (path === '/' ? '/index' : path) + '.jade');
-    if (!await fs.exists(fileName)) {
+    if (!(await fileExists(fileName))) {
       fileName = join(CONTENT_DIR, path + '/index.jade');
     }
 
-    if (!await fs.exists(fileName)) {
-      res.status(404).send({error: `The page '${path}' is not found.`});
+    if (!(await fileExists(fileName))) {
+      res.status(404).send({ error: `The page '${path}' is not found.` });
     } else {
-      const source = await fs.readFile(fileName, { encoding: 'utf8' });
+      const source = await readFile(fileName, { encoding: 'utf8' });
       const content = parseJade(path, source);
       res.status(200).send(content);
     }
@@ -46,4 +58,3 @@ router.get('/', async (req, res, next) => {
 });
 
 export default router;
-
