@@ -1,22 +1,33 @@
-/*! React Starter Kit | MIT License | http://www.reactstarterkit.com/ */
+/**
+ * React Starter Kit (https://www.reactstarterkit.com/)
+ *
+ * Copyright © 2014-2016 Kriasoft, LLC. All rights reserved.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE.txt file in the root directory of this source tree.
+ */
 
-import React, { PropTypes } from 'react';
-import styles from './ContactPage.css';
+import React, { Component, PropTypes } from 'react';
+import s from './ContactPage.scss';
 import withStyles from '../../decorators/withStyles';
 
-@withStyles(styles)
-class ContactPage {
+const title = 'Contact Us';
+
+@withStyles(s)
+class ContactPage extends Component {
 
   static contextTypes = {
-    onSetTitle: PropTypes.func.isRequired
+    onSetTitle: PropTypes.func.isRequired,
   };
 
-  render() {
-    let title = 'Contact Us';
+  componentWillMount() {
     this.context.onSetTitle(title);
+  }
+
+  render() {
     return (
-      <div className="ContactPage">
-        <div className="ContactPage-container">
+      <div className={s.root}>
+        <div className={s.container}>
           <h1>{title}</h1>
           <p>...</p>
         </div>
