@@ -1,31 +1,36 @@
-/*! React Starter Kit | MIT License | http://www.reactstarterkit.com/ */
+/**
+ * React Starter Kit (https://www.reactstarterkit.com/)
+ *
+ * Copyright © 2014-2016 Kriasoft, LLC. All rights reserved.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE.txt file in the root directory of this source tree.
+ */
 
-import React, { PropTypes } from 'react';
-import styles from './ContentPage.css';
+import React, { Component, PropTypes } from 'react';
+import s from './ContentPage.scss';
 import withStyles from '../../decorators/withStyles';
 
-@withStyles(styles)
-class ContentPage {
+@withStyles(s)
+class ContentPage extends Component {
 
   static propTypes = {
     path: PropTypes.string.isRequired,
     content: PropTypes.string.isRequired,
-    title: PropTypes.string
+    title: PropTypes.string,
   };
 
   static contextTypes = {
-    onSetTitle: PropTypes.func.isRequired
+    onSetTitle: PropTypes.func.isRequired,
   };
 
   render() {
     this.context.onSetTitle(this.props.title);
     return (
-      <div className="ContentPage">
-        <div className="ContentPage-container">
-          {
-            this.props.path === '/' ? null : <h1>{this.props.title}</h1>
-          }
-          <div dangerouslySetInnerHTML={{__html: this.props.content || ''}} />
+      <div className={s.root}>
+        <div className={s.container}>
+          {this.props.path === '/' ? null : <h1>{this.props.title}</h1>}
+          <div dangerouslySetInnerHTML={{ __html: this.props.content || '' }} />
         </div>
       </div>
     );
