@@ -1,20 +1,31 @@
-/*! React Starter Kit | MIT License | http://www.reactstarterkit.com/ */
+/**
+ * React Starter Kit (https://www.reactstarterkit.com/)
+ *
+ * Copyright © 2014-2016 Kriasoft, LLC. All rights reserved.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE.txt file in the root directory of this source tree.
+ */
 
-import React, { PropTypes } from 'react';
+import React, { Component, PropTypes } from 'react';
+import s from './ErrorPage.scss';
 import withStyles from '../../decorators/withStyles';
-import styles from './ErrorPage.css';
 
-@withStyles(styles)
-class ErrorPage {
+const title = 'Error';
+
+@withStyles(s)
+class ErrorPage extends Component {
 
   static contextTypes = {
     onSetTitle: PropTypes.func.isRequired,
-    onPageNotFound: PropTypes.func.isRequired
+    onPageNotFound: PropTypes.func.isRequired,
   };
 
-  render() {
-    let title = 'Error';
+  componentWillMount() {
     this.context.onSetTitle(title);
+  }
+
+  render() {
     return (
       <div>
         <h1>{title}</h1>
