@@ -1,19 +1,26 @@
-/*! React Starter Kit | MIT License | http://www.reactstarterkit.com/ */
+/**
+ * React Starter Kit (https://www.reactstarterkit.com/)
+ *
+ * Copyright © 2014-2016 Kriasoft, LLC. All rights reserved.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE.txt file in the root directory of this source tree.
+ */
 
-import React from 'react';
-import styles from './Feedback.css';
+import React, { Component } from 'react';
+import s from './Feedback.scss';
 import withStyles from '../../decorators/withStyles';
 
-@withStyles(styles)
-class Feedback {
+@withStyles(s)
+class Feedback extends Component {
 
   render() {
     return (
-      <div className="Feedback">
-        <div className="Feedback-container">
-          <a className="Feedback-link" href="https://gitter.im/kriasoft/react-starter-kit">Ask a question</a>
-          <span className="Feedback-spacer">|</span>
-          <a className="Feedback-link" href="https://github.com/kriasoft/react-starter-kit/issues/new">Report an issue</a>
+      <div className={s.root}>
+        <div className={s.container}>
+          <a className={s.link} href="https://gitter.im/kriasoft/react-starter-kit">Ask a question</a>
+          <span className={s.spacer}>|</span>
+          <a className={s.link} href="https://github.com/kriasoft/react-starter-kit/issues/new">Report an issue</a>
         </div>
       </div>
     );
