@@ -1,39 +1,31 @@
-/*! React Starter Kit | MIT License | http://www.reactstarterkit.com/ */
+/**
+ * React Starter Kit (https://www.reactstarterkit.com/)
+ *
+ * Copyright © 2014-2016 Kriasoft, LLC. All rights reserved.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE.txt file in the root directory of this source tree.
+ */
 
-import React, { PropTypes } from 'react';
-import styles from './Footer.css';
-import withViewport from '../../decorators/withViewport';
+import React, { Component } from 'react';
+import s from './Footer.scss';
 import withStyles from '../../decorators/withStyles';
-import Link from '../../utils/Link';
+import Link from '../Link';
 
-@withViewport
-@withStyles(styles)
-class Footer {
-
-  static propTypes = {
-    viewport: PropTypes.shape({
-      width: PropTypes.number.isRequired,
-      height: PropTypes.number.isRequired
-    }).isRequired
-  };
+@withStyles(s)
+class Footer extends Component {
 
   render() {
-    // This is just an example how one can render CSS
-    let { width, height } = this.props.viewport;
-    this.renderCss(`.Footer-viewport:after {content:' ${width}x${height}';}`);
-
     return (
-      <div className="Footer">
-        <div className="Footer-container">
-          <span className="Footer-text">© Your Company</span>
-          <span className="Footer-spacer">·</span>
-          <a className="Footer-link" href="/" onClick={Link.handleClick}>Home</a>
-          <span className="Footer-spacer">·</span>
-          <a className="Footer-link" href="/privacy" onClick={Link.handleClick}>Privacy</a>
-          <span className="Footer-spacer">·</span>
-          <a className="Footer-link" href="/not-found" onClick={Link.handleClick}>Not Found</a>
-          <span className="Footer-spacer"> | </span>
-          <span ref="viewport" className="Footer-viewport Footer-text Footer-text--muted">Viewport:</span>
+      <div className={s.root}>
+        <div className={s.container}>
+          <span className={s.text}>© Your Company</span>
+          <span className={s.spacer}>·</span>
+          <Link className={s.link} to="/">Home</Link>
+          <span className={s.spacer}>·</span>
+          <Link className={s.link} to="/privacy">Privacy</Link>
+          <span className={s.spacer}>·</span>
+          <Link className={s.link} to="/not-found">Not Found</Link>
         </div>
       </div>
     );
