@@ -1,27 +1,34 @@
-/*! React Starter Kit | MIT License | http://www.reactstarterkit.com/ */
+/**
+ * React Starter Kit (https://www.reactstarterkit.com/)
+ *
+ * Copyright © 2014-2016 Kriasoft, LLC. All rights reserved.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE.txt file in the root directory of this source tree.
+ */
 
-import React, { PropTypes } from 'react';
-import classNames from 'classnames';
-import styles from './Navigation.css';
+import React, { Component, PropTypes } from 'react';
+import cx from 'classnames';
+import s from './Navigation.scss';
 import withStyles from '../../decorators/withStyles';
-import Link from '../../utils/Link';
+import Link from '../Link';
 
-@withStyles(styles)
-class Navigation {
+@withStyles(s)
+class Navigation extends Component {
 
   static propTypes = {
-    className: PropTypes.string
+    className: PropTypes.string,
   };
 
   render() {
     return (
-      <div className={classNames(this.props.className, 'Navigation')} role="navigation">
-        <a className="Navigation-link" href="/about" onClick={Link.handleClick}>About</a>
-        <a className="Navigation-link" href="/contact" onClick={Link.handleClick}>Contact</a>
-        <span className="Navigation-spacer"> | </span>
-        <a className="Navigation-link" href="/login" onClick={Link.handleClick}>Log in</a>
-        <span className="Navigation-spacer">or</span>
-        <a className="Navigation-link Navigation-link--highlight" href="/register" onClick={Link.handleClick}>Sign up</a>
+      <div className={cx(s.root, this.props.className)} role="navigation">
+        <Link className={s.link} to="/about">About</Link>
+        <Link className={s.link} to="/contact">Contact</Link>
+        <span className={s.spacer}> | </span>
+        <Link className={s.link} to="/login">Log in</Link>
+        <span className={s.spacer}>or</span>
+        <Link className={cx(s.link, s.highlight)} to="/register">Sign up</Link>
       </div>
     );
   }
