@@ -1,21 +1,32 @@
-/*! React Starter Kit | MIT License | http://www.reactstarterkit.com/ */
+/**
+ * React Starter Kit (https://www.reactstarterkit.com/)
+ *
+ * Copyright © 2014-2016 Kriasoft, LLC. All rights reserved.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE.txt file in the root directory of this source tree.
+ */
 
-import React, { PropTypes } from 'react';
+import React, { Component, PropTypes } from 'react';
+import s from './NotFoundPage.scss';
 import withStyles from '../../decorators/withStyles';
-import styles from './NotFoundPage.css';
 
-@withStyles(styles)
-class NotFoundPage {
+const title = 'Page Not Found';
+
+@withStyles(s)
+class NotFoundPage extends Component {
 
   static contextTypes = {
     onSetTitle: PropTypes.func.isRequired,
-    onPageNotFound: PropTypes.func.isRequired
+    onPageNotFound: PropTypes.func.isRequired,
   };
 
-  render() {
-    let title = 'Page Not Found';
+  componentWillMount() {
     this.context.onSetTitle(title);
     this.context.onPageNotFound();
+  }
+
+  render() {
     return (
       <div>
         <h1>{title}</h1>
