@@ -1,26 +1,33 @@
-/*! React Starter Kit | MIT License | http://www.reactstarterkit.com/ */
+/**
+ * React Starter Kit (https://www.reactstarterkit.com/)
+ *
+ * Copyright © 2014-2016 Kriasoft, LLC. All rights reserved.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE.txt file in the root directory of this source tree.
+ */
 
-import React, { PropTypes } from 'react';
+import React, { Component, PropTypes } from 'react';
+import s from './TextBox.scss';
 import withStyles from '../../decorators/withStyles';
-import styles from './TextBox.css';
 
-@withStyles(styles)
-class TextBox {
+@withStyles(s)
+class TextBox extends Component {
 
   static propTypes = {
-    maxLines: PropTypes.number
+    maxLines: PropTypes.number,
   };
 
   static defaultProps = {
-    maxLines: 1
+    maxLines: 1,
   };
 
   render() {
     return (
-      <div className="TextBox">
+      <div className={s.root}>
         {this.props.maxLines > 1 ?
-          <textarea {...this.props} className="TextBox-input" ref="input" key="input" rows={this.props.maxLines} /> :
-          <input {...this.props} className="TextBox-input" ref="input" key="input" />}
+          <textarea {...this.props} className={s.input} ref="input" key="input" rows={this.props.maxLines} /> :
+          <input {...this.props} className={s.input} ref="input" key="input" />}
       </div>
     );
   }
