@@ -1,76 +1,31 @@
-/*! React Starter Kit | MIT License | http://www.reactstarterkit.com/ */
+/**
+ * React Starter Kit (https://www.reactstarterkit.com/)
+ *
+ * Copyright © 2014-2016 Kriasoft, LLC. All rights reserved.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE.txt file in the root directory of this source tree.
+ */
 
-import React, { PropTypes } from 'react'; // eslint-disable-line no-unused-vars
-import invariant from 'fbjs/lib/invariant';
-import { canUseDOM } from 'fbjs/lib/ExecutionEnvironment';
+import React, { Component, PropTypes } from 'react';
 
-let count = 0;
-
-function withStyles(styles) {
-  return (ComposedComponent) => class WithStyles {
-
+function withStyles(...styles) {
+  return (BaseComponent) => class StyledComponent extends Component {
     static contextTypes = {
-      onInsertCss: PropTypes.func
+      insertCss: PropTypes.func.isRequired,
     };
 
-    constructor() {
-      this.refCount = 0;
-      ComposedComponent.prototype.renderCss = function (css) {
-        let style;
-        if (canUseDOM) {
-          if (this.styleId && (style = document.getElementById(this.styleId))) {
-            if ('textContent' in style) {
-              style.textContent = css;
-            } else {
-              style.styleSheet.cssText = css;
-            }
-          } else {
-            this.styleId = `dynamic-css-${count++}`;
-            style = document.createElement('style');
-            style.setAttribute('id', this.styleId);
-            style.setAttribute('type', 'text/css');
-
-            if ('textContent' in style) {
-              style.textContent = css;
-            } else {
-              style.styleSheet.cssText = css;
-            }
-
-            document.getElementsByTagName('head')[0].appendChild(style);
-            this.refCount++;
-          }
-        } else {
-          this.context.onInsertCss(css);
-        }
-      }.bind(this);
-    }
-
     componentWillMount() {
-      if (canUseDOM) {
-        invariant(styles.use, `The style-loader must be configured with reference-counted API.`);
-        styles.use();
-      } else {
-        this.context.onInsertCss(styles.toString());
-      }
+      this.removeCss = this.context.insertCss.apply(undefined, styles);
     }
 
     componentWillUnmount() {
-      styles.unuse();
-      if (this.styleId) {
-        this.refCount--;
-        if (this.refCount < 1) {
-          let style = document.getElementById(this.styleId);
-          if (style) {
-            style.parentNode.removeChild(style);
-          }
-        }
-      }
+      this.removeCss();
     }
 
     render() {
-      return <ComposedComponent {...this.props} />;
+      return <BaseComponent {...this.props} />;
     }
-
   };
 }
 
