@@ -1,16 +1,23 @@
-/*! React Starter Kit | MIT License | http://www.reactstarterkit.com/ */
+/**
+ * React Starter Kit (https://www.reactstarterkit.com/)
+ *
+ * Copyright © 2014-2016 Kriasoft, LLC. All rights reserved.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE.txt file in the root directory of this source tree.
+ */
 
 import React, { Component } from 'react'; // eslint-disable-line no-unused-vars
 import EventEmitter from 'eventemitter3';
 import { canUseDOM } from 'fbjs/lib/ExecutionEnvironment';
 
 let EE;
-let viewport = {width: 1366, height: 768}; // Default size for server-side rendering
+let viewport = { width: 1366, height: 768 }; // Default size for server-side rendering
 const RESIZE_EVENT = 'resize';
 
 function handleWindowResize() {
   if (viewport.width !== window.innerWidth || viewport.height !== window.innerHeight) {
-    viewport = {width: window.innerWidth, height: window.innerHeight};
+    viewport = { width: window.innerWidth, height: window.innerHeight };
     EE.emit(RESIZE_EVENT, viewport);
   }
 }
@@ -22,7 +29,7 @@ function withViewport(ComposedComponent) {
       super();
 
       this.state = {
-        viewport: canUseDOM ? {width: window.innerWidth, height: window.innerHeight} : viewport
+        viewport: canUseDOM ? { width: window.innerWidth, height: window.innerHeight } : viewport,
       };
     }
 
@@ -32,6 +39,7 @@ function withViewport(ComposedComponent) {
         window.addEventListener('resize', handleWindowResize);
         window.addEventListener('orientationchange', handleWindowResize);
       }
+
       EE.on(RESIZE_EVENT, this.handleResize, this);
     }
 
@@ -49,7 +57,7 @@ function withViewport(ComposedComponent) {
     }
 
     handleResize(value) {
-      this.setState({viewport: value});
+      this.setState({ viewport: value }); // eslint-disable-line react/no-set-state
     }
 
   };
