@@ -1,20 +1,25 @@
 /**
- * React Starter Kit (http://www.reactstarterkit.com/)
+ * React Starter Kit (https://www.reactstarterkit.com/)
  *
- * Copyright © 2014-2015 Kriasoft, LLC. All rights reserved.
+ * Copyright © 2014-2016 Kriasoft, LLC. All rights reserved.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE.txt file in the root directory of this source tree.
  */
 
+import run from './run';
+import clean from './clean';
+import copy from './copy';
+import bundle from './bundle';
+
 /**
  * Compiles the project from source files into a distributable
  * format and copies it to the output (build) folder.
  */
-export default async () => {
-  console.log('build');
-  await require('./clean')();
-  await require('./copy')();
-  await require('./bundle')();
-};
+async function build() {
+  await run(clean);
+  await run(copy);
+  await run(bundle);
+}
 
+export default build;
