@@ -1,7 +1,7 @@
 /**
- * React Starter Kit (http://www.reactstarterkit.com/)
+ * React Starter Kit (https://www.reactstarterkit.com/)
  *
- * Copyright © 2014-2015 Kriasoft, LLC. All rights reserved.
+ * Copyright © 2014-2016 Kriasoft, LLC. All rights reserved.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE.txt file in the root directory of this source tree.
@@ -13,13 +13,9 @@ import fs from './lib/fs';
 /**
  * Cleans up the output (build) directory.
  */
-export default () => new Promise((resolve, reject) => {
-  console.log('clean');
-  del(['.tmp', 'build/*', '!build/.git'], {dot: true}, err => {
-    if (err) {
-      reject(err);
-    } else {
-      fs.makeDir('build/public').then(resolve, reject);
-    }
-  });
-});
+async function clean() {
+  await del(['.tmp', 'build/*', '!build/.git'], { dot: true });
+  await fs.makeDir('build/public');
+}
+
+export default clean;
