@@ -1,34 +1,28 @@
 /**
- * React Starter Kit (http://www.reactstarterkit.com/)
+ * React Starter Kit (https://www.reactstarterkit.com/)
  *
- * Copyright © 2014-2015 Kriasoft, LLC. All rights reserved.
+ * Copyright © 2014-2016 Kriasoft, LLC. All rights reserved.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE.txt file in the root directory of this source tree.
  */
 
 import path from 'path';
+import gaze from 'gaze';
 import replace from 'replace';
-import copy from './lib/copy';
-import watch from './lib/watch';
+import Promise from 'bluebird';
 
 /**
  * Copies static files such as robots.txt, favicon.ico to the
  * output (build) folder.
  */
-export default async () => {
-  console.log('copy');
+async function copy({ watch } = {}) {
+  const ncp = Promise.promisify(require('ncp'));
+
   await Promise.all([
-    // Static files
-    copy('src/public', 'build/public'),
-
-    // Files with content (e.g. *.md files)
-    copy('src/content', 'build/content'),
-
-    // Website and email templates
-    copy('src/templates', 'build/templates'),
-
-    copy('package.json', 'build/package.json')
+    ncp('src/public', 'build/public'),
+    ncp('src/content', 'build/content'),
+    ncp('package.json', 'build/package.json'),
   ]);
 
   replace({
@@ -36,14 +30,18 @@ export default async () => {
     replacement: '"start": "node server.js"',
     paths: ['build/package.json'],
     recursive: false,
-    silent: false
+    silent: false,
   });
 
-  if (global.WATCH) {
-    const watcher = await watch('src/content/**/*.*');
+  if (watch) {
+    const watcher = await new Promise((resolve, reject) => {
+      gaze('src/content/**/*.*', (err, val) => err ? reject(err) : resolve(val));
+    });
     watcher.on('changed', async (file) => {
-      file = file.substr(path.join(__dirname, '../src/content/').length);
-      await copy(`src/content/${file}`, `build/content/${file}`);
+      const relPath = file.substr(path.join(__dirname, '../src/content/').length);
+      await ncp(`src/content/${relPath}`, `build/content/${relPath}`);
     });
   }
-};
+}
+
+export default copy;
