@@ -1,7 +1,7 @@
 /**
- * React Starter Kit (http://www.reactstarterkit.com/)
+ * React Starter Kit (https://www.reactstarterkit.com/)
  *
- * Copyright © 2014-2015 Kriasoft, LLC. All rights reserved.
+ * Copyright © 2014-2016 Kriasoft, LLC. All rights reserved.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE.txt file in the root directory of this source tree.
@@ -10,24 +10,12 @@
 import fs from 'fs';
 import mkdirp from 'mkdirp';
 
-const writeFile = (filename, contents) => new Promise((resolve, reject) => {
-  fs.writeFile(filename, contents, 'utf8', err => {
-    if (err) {
-      reject(err);
-    } else {
-      resolve();
-    }
-  });
+const writeFile = (file, contents) => new Promise((resolve, reject) => {
+  fs.writeFile(file, contents, 'utf8', err => err ? reject(err) : resolve());
 });
 
-const makeDir = name => new Promise((resolve, reject) => {
-  mkdirp(name, err => {
-    if (err) {
-      reject(err);
-    } else {
-      resolve();
-    }
-  });
+const makeDir = (name) => new Promise((resolve, reject) => {
+  mkdirp(name, err => err ? reject(err) : resolve());
 });
 
 export default { writeFile, makeDir };
